// packages/lib/ren-tx/src/types/mint.ts
export type AllGatewayTransactions<DepositType> =
    | ConfirmingGatewayTransaction<DepositType>
    | AcceptedGatewayTransaction<DepositType>
    | SubmittingGatewayTransaction<DepositType>
    | MintedGatewayTransaction<DepositType>
    | CompletedGatewayTransaction<DepositType>
    | GatewayTransaction<DepositType>;

export const isConfirming = <X>(
    x: AllGatewayTransactions<X>,
): x is ConfirmingGatewayTransaction<X> => {
    return (x as ConfirmingGatewayTransaction<X>).sourceTxConfs !== undefined;
};

export interface ConfirmingGatewayTransaction<DepositType>
    extends GatewayTransaction<DepositType> {
    /**
     * How many confirmations needed to consider the source tx accepted
     */
    sourceTxConfTarget: number;
}

export const isAccepted = <X>(
    x: AllGatewayTransactions<X>,
): x is AcceptedGatewayTransaction<X> => {
    return (x as AcceptedGatewayTransaction<X>).renSignature !== undefined;
};

export interface AcceptedGatewayTransaction<DepositType>
    extends ConfirmingGatewayTransaction<DepositType> {
    /**
     * Response to renvm signing request
     */
    renResponse: LockAndMintTransaction | BurnAndReleaseTransaction;
    /**
     * Signature generated by renvm
     */
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    renSignature: any;
}

export const isSubmitted = <X>(
    x: AllGatewayTransactions<X>,
): x is SubmittingGatewayTransaction<X> => {
    return (x as SubmittingGatewayTransaction<X>).contractParams !== undefined;
};

export interface SubmittingGatewayTransaction<DepositType, CustomParams = {}>
    extends AcceptedGatewayTransaction<DepositType> {
    /**
     * Additional parameters for constructing a custom transaction
     * Provided at the point of submission
     */
    contractParams: CustomParams;
}

export const isMinted = <X>(
    x: AllGatewayTransactions<X>,
): x is MintedGatewayTransaction<X> => {
    return (x as MintedGatewayTransaction<X>).destTxHash !== undefined;
};

export interface MintedGatewayTransaction<DepositType>
    extends SubmittingGatewayTransaction<DepositType> {
    /* Hash of output transaction */
    destTxHash: string;
    /**
     * Underlying dest chain tx
     */
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    rawDestTx: any;

    /**
     * Transaction amount in the minimum denomination (eg. SATs for Bitcoin)
     */
    destTxAmount: string;
}

export const isCompleted = <X>(
    x: AllGatewayTransactions<X>,
): x is CompletedGatewayTransaction<X> => {
    return (x as CompletedGatewayTransaction<X>).completedAt !== undefined;
};

export interface CompletedGatewayTransaction<DepositType>
    extends MintedGatewayTransaction<DepositType> {
    /**
     * Unix time when mint was successful
     */
    completedAt: Number;
}

/**
 * A pair of transactions from a sourceChain to a destChain
 */
export interface GatewayTransaction<DepositType> {
    /* Hash of renVM transaction */
    renVMHash: string;
    /**
     * Transaction amount in the minimum denomination (eg. SATs for Bitcoin)
     */
    sourceTxAmount: string;
    /**
     * The RenVM Tx hash when minting,
     * The host chain burn tx hash when burning
     */
    sourceTxHash: string;
    /**
     * Current confirmations on source transaction
     */
    sourceTxConfs: number;
    /**
     * Underlying source chain tx
     */
    // eslint-disable-next-line @typescript-eslint/no-explicit-any
    rawSourceTx: DepositCommon<DepositType>;

    /**
     * Latest error detected during processing a specific transaction
     */
    error?: Error;

    /**
     * Unix time when deposit was first detected
     */
    detectedAt: Number;
}